TypeSharp.1.5.4

+++ TypeSharp/~Types/TsEnumValue.cs
@@ -3,7 +3,6 @@
     public class TsEnumValue
     {
         public string Name { get; set; }
-        public int Value { get; set; }
-
+        public long Value { get; set; }
     }
 }
